+++ src/components/todoList.tsx
@@ -1,0 +1,63 @@
+import classNames from 'classnames';
+import { TypeTodoList } from '../types/Todo';
+
+export const TodoList: React.FC<TypeTodoList> = ({
+  filteredTodoList,
+  deletingTodos,
+  handleToggleCompletion,
+  handleDeleteTodo,
+}) => {
+  return (
+    <>
+      {filteredTodoList.map(todo => (
+        <div
+          data-cy="Todo"
+          className={classNames('todo', {
+            completed: todo.completed,
+            'todo--loading': deletingTodos.includes(todo.id),
+          })}
+          key={todo.id}
+        >
+          {/* eslint-disable-next-line */}
+          <label
+            className="todo__status-label"
+            htmlFor={`todo-status-${todo.id}`}
+          >
+            <input
+              data-cy="TodoStatus"
+              type="checkbox"
+              className="todo__status"
+              checked={todo.completed}
+              id={`todo-status-${todo.id}`}
+              onChange={() => handleToggleCompletion(todo.id)}
+              disabled={deletingTodos.includes(todo.id)}
+            />
+          </label>
+
+          <span data-cy="TodoTitle" className="todo__title">
+            {todo.title}
+          </span>
+          <button
+            type="button"
+            className="todo__remove"
+            data-cy="TodoDelete"
+            onClick={() => handleDeleteTodo(todo.id)}
+            disabled={deletingTodos.includes(todo.id)}
+          >
+            ×
+          </button>
+
+          <div
+            data-cy="TodoLoader"
+            className={classNames('modal overlay', {
+              'is-active': deletingTodos.includes(todo.id),
+            })}
+          >
+            <div className="modal-background has-background-white-ter" />
+            <div className="loader" />
+          </div>
+        </div>
+      ))}
+    </>
+  );
+};
