--- src/utils/sendError.ts
@@ -1,9 +1,0 @@
-export function sendErrorMessage(
-  errorText: string,
-  setErrorMessage: React.Dispatch<React.SetStateAction<string>>,
-) {
-  setErrorMessage(errorText);
-  setTimeout(() => {
-    setErrorMessage('');
-  }, 3000);
-}
